chore: split apollo dependencies and fastify dependencies
