+++ src/lib/apollo-server.ts
@@ -1,11 +1,13 @@
 import { unwrapResolverError } from "@apollo/server/errors";
+import { Booking, Cabin, Member, Organization, Prisma, Event } from "@prisma/client";
 import { FastifyInstance, FastifyReply, FastifyRequest } from "fastify";
 import { GraphQLFormattedError } from "graphql";
 import { ZodError } from "zod";
 
+import { BookingStatus } from "@/domain/cabins.js";
 import { BaseError, codes, InternalServerError, ValidationError } from "@/domain/errors.js";
-
-import { ServiceDependencies } from "./fastify/dependencies.js";
+import { Role } from "@/domain/organizations.js";
+import { User } from "@/domain/users.js";
 
 export function getFormatErrorHandler(log?: Partial<FastifyInstance["log"]>) {
   const formatError = (formattedError: GraphQLFormattedError, error: unknown): GraphQLFormattedError => {
@@ -39,7 +41,7 @@ export function getFormatErrorHandler(log?: Partial<FastifyInstance["log"]>) {
   return formatError;
 }
 
-export interface ApolloContext extends ServiceDependencies {
+export interface ApolloContext extends ApolloServerDependencies {
   res: FastifyReply;
   req: FastifyRequest;
 }
@@ -48,4 +50,59 @@ declare module "graphql" {
   interface GraphQLErrorExtensions {
     code: string;
   }
+}
+
+interface IOrganizationService {
+  hasRole(data: { userId: string; organizationId: string; role: Role }): Promise<boolean>;
+  create(data: { name: string; description?: string; userId: string }): Promise<Organization>;
+  update(userId: string, organizationId: string, data: { name?: string; description?: string }): Promise<Organization>;
+  addMember(userId: string, data: { userId: string; organizationId: string; role: Role }): Promise<Member>;
+  removeMember(userId: string, data: { userId: string; organizationId: string } | { id: string }): Promise<Member>;
+  getMembers(userId: string, organizationId: string): Promise<Member[]>;
+  get(id: string): Promise<Organization>;
+}
+
+interface IUserService {
+  get(id: string): Promise<User>;
+  getAll(): Promise<User[]>;
+  getByFeideID(feideId: string): Promise<User | null>;
+  update(id: string, data: Prisma.UserUpdateInput): Promise<User>;
+  login(id: string): Promise<User>;
+  create(data: Prisma.UserCreateInput): Promise<User>;
+  canUpdateYear(user: Pick<User, "graduationYearUpdatedAt">): boolean;
+}
+
+export interface BookingData
+  extends Pick<
+    Prisma.BookingCreateInput,
+    "email" | "firstName" | "lastName" | "startDate" | "endDate" | "phoneNumber" | "cabinId"
+  > {}
+
+export interface ICabinService {
+  newBooking(data: BookingData): Promise<Booking>;
+  updateBookingStatus(id: string, status: BookingStatus): Promise<Booking>;
+  getCabin(id: string): Promise<Cabin>;
+}
+
+interface IEventService {
+  create(
+    userId: string,
+    organizationId: string,
+    data: {
+      name: string;
+      description?: string | null;
+      startAt: Date;
+      endAt?: Date | null;
+      location?: string | null;
+      spots?: number | null;
+      slots?: { spots: number }[] | null;
+    }
+  ): Promise<Event>;
+}
+
+export interface ApolloServerDependencies {
+  userService: IUserService;
+  organizationService: IOrganizationService;
+  cabinService: ICabinService;
+  eventService: IEventService;
 }
